fashion_MNIST_sklearn
Supervised learning using K-Neighbors algorithm from Scikit learn. Dataset contain 20000 data points each with 784 features. Pixel details of images of 9 product categories. Data can be downloaded from kaggle, link is given in the 'links-to-data.txt'

This script trains of 80% of the data and uses 20% of the data for testing purposes. Using pixel details, category of a new product is detect and category image plotted.